requirements.txt
streamlit
rdkit
requests
pillow